disk: Fix bugs in bust disk cache logic
This prevents the disk from being rebuilt when --disk-size param doesn't
change. It also rebuilds the disk when doing `podman-bootc run` without
a --disk-size parameter after previously setting the disk-size. The
comparison is made by parsing the config json, so the pretty formatting
of the config json was moved into the list command.

Signed-off-by: Chris Kyrouac <ckyrouac@redhat.com>